// src/app/core/services/model/reportDataQueryOutputDto.ts
/**
 * TWMS REST API
 * TWMS REST API
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

/**
 * ouput object for the report data query (data log entries)
 */
export interface ReportDataQueryOutputDto { 
    /**
     * row id of the entry
     */
    rowId?: number;
    /**
     * utc date of the datalog entry
     */
    dateUtc?: string;
    /**
     * name of the device point
     */
    devicePointName?: string;
    /**
     * uuid of the device point
     */
    devicePointUuid?: string;
    /**
     * name of the device
     */
    deviceName?: string;
    /**
     * name of the device point, e.g. \"Beate 1\"
     */
    deviceUuid?: string;
    /**
     * type of the device, e.g. \"Wall-E\"
     */
    deviceTypeName?: string;
    /**
     * number of the channel, e.g. 6 (6 = Temperature_Get, see enum ChannelValueCodes or emico c api)
     */
    dataPointId?: string;
    /**
     * value of the data log entry, e.g. 18°C
     */
    value?: string;
    /**
     * datapoint Unit
     */
    unit?: string;
    /**
     * human-readable channel name  datapoint description
     */
    description?: string;
}